Rename Mutex and Semaphore classes
