+++ lib/RedisMutex.js
