+++ lib/RedisSemaphore.js
@@ -1,4 +1,4 @@
-const RedisMutex = require('./Mutex')
+const RedisMutex = require('./RedisMutex')
 const debug = require('debug')('redis-semaphore:semaphore:instance')
 const LostLockError = require('./errors/LostLockError')
 const semaphore = require('./semaphore')
